更新代码和数据库

+++ Uzumzki-project/2018/ball/src/main/java/com/xiaoyi/ssm/controller/venue/ApiVenueTemplateController.java
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -15,9 +14,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.xiaoyi.ssm.dto.ApiMessage;
+import com.xiaoyi.ssm.model.Field;
 import com.xiaoyi.ssm.model.Member;
 import com.xiaoyi.ssm.model.Venue;
 import com.xiaoyi.ssm.model.VenueTemplate;
+import com.xiaoyi.ssm.service.FieldService;
+import com.xiaoyi.ssm.service.FieldTemplateService;
 import com.xiaoyi.ssm.service.VenueService;
 import com.xiaoyi.ssm.service.VenueTemplateService;
 import com.xiaoyi.ssm.util.Global;
@@ -37,6 +39,10 @@ public class ApiVenueTemplateController {
 	private VenueTemplateService venueTemplateService;
 	@Autowired
 	private VenueService venueService;
+	@Autowired
+	private FieldTemplateService fieldTemplateService;
+	@Autowired
+	private FieldService fieldService;
 
 	/**  
 	 * @Description: 模板列表
@@ -58,7 +64,12 @@ public class ApiVenueTemplateController {
 			map.put("id", venueTemplate.getId());// id
 			map.put("name", venueTemplate.getName());// 模板名称
 			map.put("sysFlag", venueTemplate.getId().equals(venueTemplate.getVenueid()));// true为系统默认false不是
-			map.put("defaultFlag", venueTemplate.getDefaultflag());// 是否为默认模板(0=否1=是)
+			List<Field> fieldList = fieldService.selectByDefaultVenue(id, venueTemplate.getId());
+			String[] fields = new String[fieldList.size()];
+			for (int i = 0; i < fieldList.size(); i++) {
+				fields[i] = fieldList.get(i).getName();
+			}
+			map.put("fields", fields);// 默认使用场地名
 			listmap.add(map);
 		}
 		return new ApiMessage(200, "查询成功", listmap);
@@ -125,22 +136,31 @@ public class ApiVenueTemplateController {
 	@ResponseBody
 	public ApiMessage deleteTmplate(String id) {
 		VenueTemplate venueTemplate = venueTemplateService.selectByPrimaryKey(id);
+		if (venueTemplate == null ) {
+			return new ApiMessage(400, "删除失败,该模板不存在");
+		}
+		
 		if (id.equals(venueTemplate.getVenueid())) {
 			return new ApiMessage(400, "删除失败,系统模板不能删除");
 		}
 		
-		if (venueTemplate.getDefaultflag() == 1) {
-			// 现有默认模板删除后,将系统模板设为默认
-			VenueTemplate sysVenueTemplate = venueTemplateService.selectByVenueTemplate(venueTemplate.getVenueid(), venueTemplate.getVenueid());
-			sysVenueTemplate.setDefaultflag(1);
-			venueTemplateService.updateByPrimaryKeySelective(sysVenueTemplate);
+		List<Field> fieldList = fieldService.selectByDefaultVenue(id, venueTemplate.getId());
+		if (fieldList.size() > 0 ) {
+			String fields = "";
+			for (int i = 0; i < fieldList.size(); i++) {
+				fields += fieldList.get(i).getName();
+				if (i != fieldList.size() - 1) {
+					fields += ",";
+				}
+			}
+			return new ApiMessage(400, "删除失败,请先把"+fields+"场地默认模板设置为其他的模板");
+		} else {
+			int flag = venueTemplateService.deleteByPrimaryKey(id);
+			if (flag > 0 ) {
+				return new ApiMessage(200, "删除成功");
+			}
+			return new ApiMessage(400, "删除失败");
 		}
-		
-		int flag = venueTemplateService.deleteByPrimaryKey(id);
-		if (flag > 0 ) {
-			return new ApiMessage(200, "删除成功");
-		}
-		return new ApiMessage(400, "删除失败");
 	}
 	
 	/**  
@@ -160,10 +180,9 @@ public class ApiVenueTemplateController {
 		if (venue == null) {
 			return new ApiMessage(400, "场馆不存在");
 		}
-		
-		if (defaultFlag == 1) {
-			//将所有模板设置为非默认
-			venueTemplateService.updateNoDefaultVenue(venue.getId());
+		List<VenueTemplate> listTemplates = venueTemplateService.selectByVenue(id);
+		if (listTemplates.size() >= 6) {
+			return new ApiMessage(400, "最多允许设置6个模板,请删除不用模板后重试");
 		}
 		
 		VenueTemplate venueTemplate = new VenueTemplate();
@@ -173,7 +192,6 @@ public class ApiVenueTemplateController {
 		venueTemplate.setCreatetime(new Date());
 		venueTemplate.setModifytime(new Date());
 		venueTemplate.setVenueid(venue.getId());
-		venueTemplate.setDefaultflag(defaultFlag);
 		
 		int flag = venueTemplateService.insertSelective(venueTemplate);
 		if (flag > 0 ) {
@@ -192,7 +210,7 @@ public class ApiVenueTemplateController {
 	 */ 
 	@RequestMapping(value = "/setDefaultTmplate")
 	@ResponseBody
-	public ApiMessage setDefaultTmplate(String id, String venueId, HttpServletRequest request) {
+	public ApiMessage setDefaultTmplate(String id, String venueId, HttpServletRequest request, String fieldStr) {
 
 		String token = (String) request.getAttribute("token");
 		Member member = (Member) RedisUtil.getRedisOne(Global.redis_member, token);
@@ -203,18 +221,14 @@ public class ApiVenueTemplateController {
 			return new ApiMessage(400, "场馆不存在");
 		}
 		
-		//将所有模板设置为非默认
-		venueTemplateService.updateNoDefaultVenue(venue.getId());
-		
-		//设置唯一默认模板
-		VenueTemplate venueTemplate = venueTemplateService.selectByPrimaryKey(id);
-		venueTemplate.setDefaultflag(1);
-		
-		int flag = venueTemplateService.updateByPrimaryKeySelective(venueTemplate);
-		if (flag > 0 ) {
-			return new ApiMessage(200, "设置成功");
+		// 解析场地ID设置为默认
+		String[] fields = fieldStr.split(",");
+		for (int i = 0; i < fields.length; i++) {
+			Field field = fieldService.selectByPrimaryKey(fields[i]);
+			field.setDefaultTemplate(id);
+			fieldService.updateByPrimaryKeySelective(field);
 		}
-		return new ApiMessage(400, "设置失败");
+		return new ApiMessage(200, "修改成功");
 	}
 	
 }
